+++ www/run/tulipcc-20250302084240.mjs
@@ -69,8 +69,8 @@ if (ENVIRONMENT_IS_NODE) {
         // web worker
         PACKAGE_PATH = encodeURIComponent(location.pathname.substring(0, location.pathname.lastIndexOf('/')) + '/');
       }
-      var PACKAGE_NAME = 'build-standard/tulip/obj/tulipcc-20250302081851.data';
-      var REMOTE_PACKAGE_BASE = 'tulipcc-20250302081851.data';
+      var PACKAGE_NAME = 'build-standard/tulip/obj/tulipcc-20250302084240.data';
+      var REMOTE_PACKAGE_BASE = 'tulipcc-20250302084240.data';
       var REMOTE_PACKAGE_NAME = Module['locateFile'] ? Module['locateFile'](REMOTE_PACKAGE_BASE, '') : REMOTE_PACKAGE_BASE;
 var REMOTE_PACKAGE_SIZE = metadata['remote_package_size'];
 
@@ -213,10 +213,10 @@ Module['FS_createPath']("/tulip4/sys/im", "tiny_town", true, true);
           var files = metadata['files'];
           for (var i = 0; i < files.length; ++i) {
             DataRequest.prototype.requests[files[i].filename].onload();
-          }          Module['removeRunDependency']('datafile_build-standard/tulip/obj/tulipcc-20250302081851.data');
+          }          Module['removeRunDependency']('datafile_build-standard/tulip/obj/tulipcc-20250302084240.data');
 
       };
-      Module['addRunDependency']('datafile_build-standard/tulip/obj/tulipcc-20250302081851.data');
+      Module['addRunDependency']('datafile_build-standard/tulip/obj/tulipcc-20250302084240.data');
 
       Module['preloadResults'] ??= {};
 
@@ -930,14 +930,14 @@ function createExportWrapper(name, nargs) {
 var wasmBinaryFile;
 function findWasmBinary() {
   if (Module['locateFile']) {
-    var f = 'tulipcc-20250302081851.wasm';
+    var f = 'tulipcc-20250302084240.wasm';
     if (!isDataURI(f)) {
       return locateFile(f);
     }
     return f;
   }
   // Use bundler-friendly `new URL(..., import.meta.url)` pattern; works in browsers too.
-  return new URL('tulipcc-20250302081851.wasm', import.meta.url).href;
+  return new URL('tulipcc-20250302084240.wasm', import.meta.url).href;
 }
 
 function getBinarySync(file) {
@@ -11513,7 +11513,7 @@ for (const prop of Object.keys(Module)) {
 })();
 export default _createMicroPythonModule;
 /*
- * This file is part of the MicroPython project, http://tulipcc-20250302081851.org/
+ * This file is part of the MicroPython project, http://tulipcc-20250302084240.org/
  *
  * The MIT License (MIT)
  *
@@ -11541,7 +11541,7 @@ export default _createMicroPythonModule;
 // Options:
 // - pystack: size in words of the MicroPython Python stack.
 // - heapsize: size in bytes of the MicroPython GC heap.
-// - url: location to load `tulipcc-20250302081851.mjs`.
+// - url: location to load `tulipcc-20250302084240.mjs`.
 // - stdin: function to return input characters.
 // - stdout: function that takes one argument, and is passed lines of stdout
 //   output as they are produced.  By default this is handled by Emscripten
@@ -11902,7 +11902,7 @@ if (
     typeof process.versions === "object" &&
     typeof process.versions.node === "string"
 ) {
-    // Check if this module is run from the command line via `node tulipcc-20250302081851.mjs`.
+    // Check if this module is run from the command line via `node tulipcc-20250302084240.mjs`.
     //
     // See https://stackoverflow.com/questions/6398196/detect-if-called-through-require-or-directly-by-command-line/66309132#66309132
     //
@@ -11926,7 +11926,7 @@ if (
 }
 
 /*
- * This file is part of the MicroPython project, http://tulipcc-20250302081851.org/
+ * This file is part of the MicroPython project, http://tulipcc-20250302084240.org/
  *
  * The MIT License (MIT)
  *
@@ -12167,7 +12167,7 @@ class PyProxyThenable {
     }
 }
 /*
- * This file is part of the MicroPython project, http://tulipcc-20250302081851.org/
+ * This file is part of the MicroPython project, http://tulipcc-20250302084240.org/
  *
  * The MIT License (MIT)
  *
